[ADD] signUp viewModel add
ADD - 회원가입 viewModel 추가, 코드 수정
